move main/validation to shared
