+++ src/main/app.ts
@@ -16,13 +16,11 @@ import * as v from 'valibot'
 
 import type { NewClientMessage } from '../services/core.js'
 import type { AppConfig, SentryEnvironment } from '../shared/app.js'
+import { FilesSelectParamsSchema } from '../shared/ipc.js'
 import type { ConfigStore } from './config-store.js'
 import { Intl } from './intl.js'
 import { setUpMainIPC } from './ipc.js'
-import {
-	FilesSelectParamsSchema,
-	ServiceErrorMessageSchema,
-} from './validation.js'
+import { ServiceErrorMessageSchema } from './service-error.js'
 
 const log = debug('comapeo:main:app')
