+++ src/preload/runtime.ts
@@ -1,15 +1,8 @@
 import type { Systeminformation } from 'systeminformation'
 
-import type {
-	PersistedCoordinateFormat,
-	PersistedLocale,
-} from '../main/validation.js'
-import type { LocaleState } from '../shared/intl.js'
-
-export type SelectedFile = {
-	name: string
-	path: string
-}
+import type { CoordinateFormat } from '../shared/coordinate-format.js'
+import type { Locale, LocaleState } from '../shared/intl.js'
+import type { SelectedFile } from '../shared/ipc.ts'
 
 export type RuntimeApi = {
 	// Files
@@ -30,15 +23,15 @@ export type RuntimeApi = {
 
 	// Settings (get)
 	getActiveProjectId: () => Promise<string | undefined>
-	getCoordinateFormat: () => Promise<PersistedCoordinateFormat>
+	getCoordinateFormat: () => Promise<CoordinateFormat>
 	getDiagnosticsEnabled: () => Promise<boolean>
 	getLocaleState: () => Promise<LocaleState>
 
 	// Settings (set)
 	setActiveProjectId: (value: string | null) => Promise<void>
-	setCoordinateFormat: (value: PersistedCoordinateFormat) => Promise<void>
+	setCoordinateFormat: (value: CoordinateFormat) => Promise<void>
 	setDiagnosticsEnabled: (value: boolean) => Promise<void>
-	setLocale: (value: PersistedLocale) => Promise<void>
+	setLocale: (value: Locale) => Promise<void>
 
 	// Sentry
 	getSentryConfig: () => {
