+++ src/renderer/src/lib/coordinate-format.test.ts
@@ -1,24 +1,6 @@
-import { is } from 'valibot'
-import { describe, expect, it, test } from 'vitest'
+import { describe, expect, it } from 'vitest'
 
-import { CoordinateFormatSchema, convertToUTM } from './coordinate-format'
-
-describe('CoordinateFormatSchema', () => {
-	test('captures valid values', () => {
-		expect(is(CoordinateFormatSchema, 'dms')).toBe(true)
-		expect(is(CoordinateFormatSchema, 'dd')).toBe(true)
-		expect(is(CoordinateFormatSchema, 'utm')).toBe(true)
-	})
-
-	test('denies invalid values', () => {
-		expect(is(CoordinateFormatSchema, '')).toBe(false)
-		expect(is(CoordinateFormatSchema, 0)).toBe(false)
-		expect(is(CoordinateFormatSchema, null)).toBe(false)
-		expect(is(CoordinateFormatSchema, undefined)).toBe(false)
-		expect(is(CoordinateFormatSchema, {})).toBe(false)
-		expect(is(CoordinateFormatSchema, [])).toBe(false)
-	})
-})
+import { convertToUTM } from './coordinate-format'
 
 describe('convertToUTM()', () => {
 	it('converts latitude + longitude to UTM format', () => {
