+++ src/renderer/src/routes/app/settings/coordinate-system.tsx
@@ -16,13 +16,11 @@ import { createFileRoute, useRouter } from '@tanstack/react-router'
 import { defineMessages, useIntl } from 'react-intl'
 import { parse } from 'valibot'
 
+import { CoordinateFormatSchema } from '../../../../../shared/coordinate-format'
 import { BLUE_GREY, DARK_GREY } from '../../../colors'
 import { ErrorDialog } from '../../../components/error-dialog'
 import { Icon } from '../../../components/icon'
-import {
-	CoordinateFormatSchema,
-	formatCoords,
-} from '../../../lib/coordinate-format'
+import { formatCoords } from '../../../lib/coordinate-format'
 import {
 	getCoordinateFormatQueryOptions,
 	setCoordinateFormatMutationOptions,
