+++ src/shared/app.ts
@@ -1,3 +1,5 @@
+import * as v from 'valibot'
+
 export type SentryEnvironment = 'development' | 'qa' | 'production'
 
 export type AppType =
@@ -6,15 +8,22 @@ export type AppType =
 	| 'release-candidate'
 	| 'production'
 
-export type AppConfig = {
+export const AppConfigSchema = v.object({
 	/** Enables ASAR format */
-	asar?: boolean
+	asar: v.optional(v.boolean()),
 	/** Sets the online map style for @comapeo/core to use */
-	onlineStyleUrl?: string
+	onlineStyleUrl: v.optional(v.pipe(v.string(), v.url())),
 	/** Sets the user data directory for the application to use */
-	userDataPath?: string
+	userDataPath: v.optional(v.string()),
 	/** Indicates the app type */
-	appType: AppType
+	appType: v.union([
+		v.literal('development'),
+		v.literal('internal'),
+		v.literal('release-candidate'),
+		v.literal('production'),
+	]),
 	/** Indicates the app version */
-	appVersion: string
-}
+	appVersion: v.string(),
+})
+
+export type AppConfig = v.InferOutput<typeof AppConfigSchema>
