+++ src/shared/intl.ts
@@ -1,3 +1,5 @@
+import * as v from 'valibot'
+
 import SUPPORTED_LANGUAGES from '../../languages.json' with { type: 'json' }
 
 export type LocaleSource = 'selected' | 'system' | 'fallback'
@@ -8,3 +10,16 @@ export type LocaleState = {
 	source: LocaleSource
 	value: SupportedLanguageTag
 }
+
+export const LocaleSchema = v.variant('useSystemPreferences', [
+	v.object({
+		useSystemPreferences: v.literal(true),
+		languageTag: v.null(),
+	}),
+	v.object({
+		useSystemPreferences: v.literal(false),
+		languageTag: v.string(),
+	}),
+])
+
+export type Locale = v.InferOutput<typeof LocaleSchema>
