+++ src/shared/ipc.ts
@@ -1,0 +1,13 @@
+import * as v from 'valibot'
+
+export type SelectedFile = {
+	name: string
+	path: string
+}
+
+export const FilesSelectParamsSchema = v.union([
+	v.object({ extensionFilters: v.optional(v.array(v.string())) }),
+	v.undefined(),
+])
+
+export type FilesSelectParams = v.InferInput<typeof FilesSelectParamsSchema>
