+++ src/shared/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": ["../../tsconfig.json"],
+	"compilerOptions": {
+		"module": "NodeNext",
+		"erasableSyntaxOnly": true
+	},
+	"include": ["**/*.js", "**/*.ts"]
+}
